requirements.txt
# Giulia v1 – pinned for reproducible installs; bump with care
chainlit==2.0.0
faiss-cpu==1.8.0.post1
mistralai==1.2.0
pymongo==4.6.3
pymupdf==1.24.9
python-docx==1.2.0
python-dotenv==1.0.1
numpy==1.26.4
